requirements.txt
streamlit
scikit-surprise
scikit-learn
pandas
numpy==1.26.4
requests
